Move functional test to unit test for create_tags
